+++ crates/polars-stream/src/nodes/io_sinks2/writers/parquet/mod.rs
@@ -1,0 +1,144 @@
+use std::sync::Arc;
+
+use arrow::datatypes::ArrowSchemaRef;
+use polars_error::PolarsResult;
+use polars_io::pl_async;
+use polars_io::prelude::{ParquetWriteOptions, get_column_write_options};
+use polars_io::utils::sync_on_close::SyncOnCloseType;
+use polars_parquet::write::{
+    ColumnWriteOptions, CompressedPage, SchemaDescriptor, Version, WriteOptions, to_parquet_schema,
+};
+use polars_utils::IdxSize;
+
+use crate::async_executor::{self, TaskPriority};
+use crate::async_primitives::connector;
+use crate::nodes::io_sinks2::components::sink_morsel::{SinkMorsel, SinkMorselPermit};
+use crate::nodes::io_sinks2::components::size::RowCountAndSize;
+use crate::nodes::io_sinks2::writers::interface::{
+    FileWriterStarter, default_ideal_sink_morsel_size,
+};
+use crate::utils::tokio_handle_ext;
+
+mod io_writer;
+mod row_group_encoder;
+
+pub struct ParquetWriterStarter {
+    pub options: ParquetWriteOptions,
+    pub arrow_schema: ArrowSchemaRef,
+    pub initialized_state: std::sync::Mutex<Option<InitializedState>>,
+    pub pipeline_depth: usize,
+    pub sync_on_close: SyncOnCloseType,
+    pub row_group_size: Option<IdxSize>,
+}
+
+#[derive(Clone)]
+pub struct InitializedState {
+    column_options: Arc<Vec<ColumnWriteOptions>>,
+    schema_descriptor: Arc<SchemaDescriptor>,
+}
+
+struct EncodedRowGroup {
+    data: Vec<Vec<CompressedPage>>,
+    morsel_permit: SinkMorselPermit,
+}
+
+impl FileWriterStarter for ParquetWriterStarter {
+    fn writer_name(&self) -> &str {
+        "parquet"
+    }
+
+    fn ideal_morsel_size(&self) -> RowCountAndSize {
+        if let Some(row_group_size) = self.row_group_size {
+            RowCountAndSize {
+                num_rows: row_group_size,
+                num_bytes: u64::MAX,
+            }
+        } else {
+            default_ideal_sink_morsel_size()
+        }
+    }
+
+    fn start_file_writer(
+        &self,
+        mut morsel_rx: connector::Receiver<SinkMorsel>,
+        file: tokio_handle_ext::AbortOnDropHandle<
+            PolarsResult<polars_io::prelude::file::Writeable>,
+        >,
+    ) -> PolarsResult<async_executor::JoinHandle<PolarsResult<()>>> {
+        let InitializedState {
+            column_options,
+            schema_descriptor,
+        } = {
+            let mut initialized_state = self.initialized_state.lock().unwrap();
+
+            if initialized_state.is_none() {
+                let column_options: Arc<Vec<ColumnWriteOptions>> = Arc::new(
+                    get_column_write_options(&self.arrow_schema, &self.options.field_overwrites),
+                );
+                let schema_descriptor = Arc::new(to_parquet_schema(
+                    &self.arrow_schema,
+                    column_options.as_ref(),
+                )?);
+
+                *initialized_state = Some(InitializedState {
+                    column_options,
+                    schema_descriptor,
+                })
+            };
+
+            initialized_state.clone().unwrap()
+        };
+
+        let (encoded_row_group_tx, encoded_row_group_rx) = tokio::sync::mpsc::channel::<
+            async_executor::AbortOnDropHandle<PolarsResult<EncodedRowGroup>>,
+        >(self.pipeline_depth);
+
+        let key_value_metadata = self.options.key_value_metadata.clone();
+        let write_options = WriteOptions {
+            statistics: self.options.statistics,
+            compression: self.options.compression.into(),
+            version: Version::V1,
+            data_page_size: self.options.data_page_size,
+        };
+
+        let sync_on_close = self.sync_on_close;
+        let arrow_schema = Arc::clone(&self.arrow_schema);
+        let num_leaf_columns = schema_descriptor.leaves().len();
+
+        let io_handle = tokio_handle_ext::AbortOnDropHandle(
+            pl_async::get_runtime().spawn(
+                io_writer::IOWriter {
+                    file,
+                    encoded_row_group_rx,
+                    arrow_schema,
+                    schema_descriptor: Arc::clone(&schema_descriptor),
+                    write_options,
+                    column_options: Arc::clone(&column_options),
+                    key_value_metadata,
+                    num_leaf_columns,
+                    sync_on_close,
+                }
+                .run(),
+            ),
+        );
+
+        let compute_handle = async_executor::AbortOnDropHandle::new(async_executor::spawn(
+            TaskPriority::High,
+            row_group_encoder::RowGroupEncoder {
+                morsel_rx,
+                encoded_row_group_tx,
+                schema_descriptor,
+                write_options,
+                column_options,
+                num_leaf_columns,
+            }
+            .run(),
+        ));
+
+        Ok(async_executor::spawn(TaskPriority::Low, async move {
+            compute_handle.await?;
+            io_handle.await.unwrap()?;
+            Ok(())
+        }))
+    }
+}
